Change all int64 to int, except Artifact.ExpiresAt

+++ activity.go
@@ -25,13 +25,13 @@ type Activity struct {
 	LinkedinArticleUrn string           `json:"linkedinArticleUrn,omitempty"`
 	ContentText        *Text            `json:"contentText,omitempty"`
 	Title              string           `json:"title,omitempty"`
-	NumLikes           int64            `json:"numLikes,omitempty"`
+	NumLikes           int              `json:"numLikes,omitempty"`
 	PostedDate         *Date            `json:"postedDate,omitempty"`
 	SocialDetail       *SocialDetail    `json:"socialDetail,omitempty"`
 	TrackingData       *TrackingData    `json:"trackingData,omitempty"`
-	NumComments        int64            `json:"numComments,omitempty"`
+	NumComments        int              `json:"numComments,omitempty"`
 	CreatedDate        *Date            `json:"createdDate,omitempty"`
-	PostedAt           int64            `json:"postedAt,omitempty"`
+	PostedAt           int              `json:"postedAt,omitempty"`
 	EntityUrn          string           `json:"entityUrn,omitempty"`
 	AuthorComponent    *AuthorComponent `json:"authorComponent,omitempty"`
 	Actor              Actor            `json:"actor,omitempty"`
@@ -68,7 +68,7 @@ type COMLinkedinVoyagerFeedRenderArticleComponent struct {
 type SubscribeAction struct {
 	EntityUrn       string `json:"entityUrn,omitempty"`
 	Subscribed      bool   `json:"subscribed,omitempty"`
-	SubscriberCount int64  `json:"subscriberCount,omitempty"`
+	SubscriberCount int    `json:"subscriberCount,omitempty"`
 }
 
 type COMLinkedinVoyagerFeedRenderArticleComponentFollowAction struct {
@@ -139,7 +139,7 @@ type SocialDetail struct {
 	SocialPermissions         *SocialPermissions         `json:"socialPermissions,omitempty"`
 	Liked                     bool                       `json:"liked,omitempty"`
 	ShowShareButton           bool                       `json:"showShareButton,omitempty"`
-	TotalShares               int64                      `json:"totalShares,omitempty"`
+	TotalShares               int                        `json:"totalShares,omitempty"`
 	Urn                       string                     `json:"urn,omitempty"`
 	ThreadID                  string                     `json:"threadId,omitempty"`
 	AllowedCommentersScope    string                     `json:"allowedCommentersScope,omitempty"`
@@ -166,16 +166,16 @@ type SocialPermissions struct {
 type TotalSocialActivityCounts struct {
 	SocialDetailEntityUrn string              `json:"socialDetailEntityUrn,omitempty"`
 	Urn                   string              `json:"urn,omitempty"`
-	NumComments           int64               `json:"numComments,omitempty"`
+	NumComments           int                 `json:"numComments,omitempty"`
 	ReactionTypeCounts    []ReactionTypeCount `json:"reactionTypeCounts,omitempty"`
 	EntityUrn             string              `json:"entityUrn,omitempty"`
-	NumShares             int64               `json:"numShares,omitempty"`
-	NumLikes              int64               `json:"numLikes,omitempty"`
+	NumShares             int                 `json:"numShares,omitempty"`
+	NumLikes              int                 `json:"numLikes,omitempty"`
 	Liked                 bool                `json:"liked,omitempty"`
 }
 
 type ReactionTypeCount struct {
-	Count        int64  `json:"count,omitempty"`
+	Count        int    `json:"count,omitempty"`
 	ReactionType string `json:"reactionType,omitempty"`
 }
